refactor: split sleeper_tools monolith — extract enrichment layer
Move the ~1000-line enrichment/data-fetch layer (schedule/snaps/injuries/practice/weekly-usage fetchers + usage-opponent enrichment helpers + ADVANCED_ENRICH_ENABLED) out of the 3022-line sleeper_tools.py into a new nfl_mcp/sleeper_enrichment.py. It's a clean leaf (public Sleeper tools call into it, not the reverse), so extraction is cycle-free. sleeper_tools.py re-exports every moved name (facade), so all sleeper_tools.<name> references and imports across the codebase are unchanged. sleeper_tools.py: 3022 -> 2012 lines. Behavior-preserving; full suite 901 passed, 2 skipped, ruff clean. Repointed 4 white-box tests (field-mappings, routes-zero) that monkeypatched create_http_client/ADVANCED_ENRICH_ENABLED to target the enrichment module (functions now resolve those globals there). Strategic-planning + core league/roster/transactions splits are follow-ups.
