+++ tests/test_sleeper_field_mappings.py
@@ -50,8 +50,8 @@ class TestSleeperFieldMappings:
 
         # Import after setting env var
         from nfl_mcp import sleeper_tools
-        monkeypatch.setattr("nfl_mcp.sleeper_tools.ADVANCED_ENRICH_ENABLED", True)
-        monkeypatch.setattr("nfl_mcp.sleeper_tools.create_http_client", mock_create_client)
+        monkeypatch.setattr("nfl_mcp.sleeper_enrichment.ADVANCED_ENRICH_ENABLED", True)
+        monkeypatch.setattr("nfl_mcp.sleeper_enrichment.create_http_client", mock_create_client)
 
         # Mock validation to always pass - it's imported inside the function
         def mock_validate(data, validator, name, allow_partial=True):
@@ -131,8 +131,8 @@ class TestSleeperFieldMappings:
 
         # Import after setting env var
         from nfl_mcp import sleeper_tools
-        monkeypatch.setattr("nfl_mcp.sleeper_tools.ADVANCED_ENRICH_ENABLED", True)
-        monkeypatch.setattr("nfl_mcp.sleeper_tools.create_http_client", mock_create_client)
+        monkeypatch.setattr("nfl_mcp.sleeper_enrichment.ADVANCED_ENRICH_ENABLED", True)
+        monkeypatch.setattr("nfl_mcp.sleeper_enrichment.create_http_client", mock_create_client)
 
         # Mock validation - it's imported inside the function
         def mock_validate(data, validator, name, allow_partial=True):
